- Replace getOpSimField function by a utility function on sims_survey_fields.  - Move utility functions treexyz, rad_lenght and buildTree to sims_utils
